pyproject.toml
[tool.poetry]
name = "geo_treker"
version = "0.1.0"
description = "API for recive geo data from user"
authors = ["Pavel <pbekrenev@sdvor.com>"]

[tool.poetry.dependencies]
python = "~3.9"
aiohttp = "^3.8.1"
aioredis = "^2.0.1"
aiopg = "^1.3.3"
uvicorn = "^0.17.6"
starlette-apispec = "^1.0.5"
fastapi = "^0.78.0"
pydantic = "^1.9.1"
uvloop = "^0.16.0"
SQLAlchemy = "^1.4.36"
GeoAlchemy2 = "^0.11.1"
alembic = "^1.7.7"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
mypy = "^0.950"
isort = "^5.10.1"
flake8 = "^4.0.1"
autoflake = "^1.4"
pre-commit = "^2.19.0"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
